Move UbloxFirmware class to its own files.
