+++ ublox_gps/include/ublox_gps/node.hpp
@@ -57,6 +57,7 @@
 #include <ublox_gps/rtcm.hpp>
 #include <ublox_gps/utils.hpp>
 #include <ublox_gps/raw_data_pa.hpp>
+#include <ublox_gps/ublox_firmware.hpp>
 #include <ublox_gps/ublox_topic_diagnostic.hpp>
 
 // This file also declares UbloxNode which implements ComponentInterface and is
@@ -290,38 +291,6 @@ class UbloxNode final {
   //! Node Handle for GPS node
   std::shared_ptr<ros::NodeHandle> nh_;
 
-};
-
-/**
- * @brief This abstract class represents a firmware component.
- *
- * @details The Firmware components update the fix diagnostics.
- */
-class UbloxFirmware : public virtual ComponentInterface {
- public:
-  //! Subscribe Rate for u-blox SV Info messages
-  constexpr static uint32_t kNavSvInfoSubscribeRate = 20;
-
-  explicit UbloxFirmware(std::shared_ptr<diagnostic_updater::Updater> updater, std::shared_ptr<Gnss> gnss, ros::NodeHandle* node);
-
-  /**
-   * @brief Add the fix diagnostics to the updater.
-   */
-  void initializeRosDiagnostics() override;
-
- protected:
-  /**
-   * @brief Handle to send fix status to ROS diagnostics.
-   */
-  virtual void fixDiagnostic(
-      diagnostic_updater::DiagnosticStatusWrapper& stat) = 0;
-
-  std::shared_ptr<diagnostic_updater::Updater> updater_;
-  std::shared_ptr<Gnss> gnss_;
-  //! The fix status service type, set in the Firmware Component
-  //! based on the enabled GNSS
-  int fix_status_service_{0};
-  ros::NodeHandle* node_;
 };
 
 /**
